Merge in new comms code

+++ state_machine/state_machine_remote.h
@@ -48,7 +48,7 @@ struct queue_remote_t {
 /* Global variables --------------------------------------------------------*/
 
 extern remote_event_t remote_state;
-extern uint32_t sub_status;
+extern uint8_t sub_status;
 
 /* MACROS for everyone--------------------------------------------------------*/
